Make filebrowsers load stored directories
